add validation api & logging

+++ src/main/java/org/nutricraft/Services/CoinServicesImpl.java
@@ -2,20 +2,32 @@ package org.nutricraft.Services;
 
 import org.nutricraft.Database.Database;
 import org.nutricraft.Model.Coins;
+import org.nutricraft.Model.LogModel;
 
+import javax.annotation.Resource;
 import javax.jws.WebMethod;
 import javax.jws.WebService;
+import javax.xml.ws.WebServiceContext;
+import javax.xml.ws.handler.MessageContext;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.Statement;
+import java.sql.Timestamp;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @WebService(endpointInterface = "org.nutricraft.Services.CoinServices")
 public class CoinServicesImpl implements CoinServices{
 
+    @Resource
+    private WebServiceContext wsContext;
     @WebMethod
     public List<Coins> getAllCoins() {
+        if(!validateApiKey()){
+            System.out.println("API KEY INVALID");
+            return null;
+        }
         List<Coins> coins = new ArrayList<Coins>();
         try{
             Database db = new Database();
@@ -26,6 +38,7 @@ public class CoinServicesImpl implements CoinServices{
             while (result.next()) {
                 coins.add(new Coins(result.getString("id"), result.getInt("coin")));
             }
+            log("Get All Coins");
         }catch (Exception e){
             e.printStackTrace();
         }
@@ -39,6 +52,10 @@ public class CoinServicesImpl implements CoinServices{
 
     @WebMethod
     public Integer getCoins(String id) {
+        if(!validateApiKey()){
+            System.out.println("API KEY INVALID");
+            return null;
+        }
         int coin=0;
         try {
             Database db = new Database();
@@ -49,6 +66,7 @@ public class CoinServicesImpl implements CoinServices{
             if(result.next()){
                 coin = result.getInt("coin");
             }
+            log("Get Coins");
         }catch (Exception e){
             e.printStackTrace();
         }
@@ -57,12 +75,17 @@ public class CoinServicesImpl implements CoinServices{
 
     @WebMethod
     public String addCoins(String id, int coins) {
+        if(!validateApiKey()){
+            System.out.println("API KEY INVALID");
+            return null;
+        }
         try{
             Database db = new Database();
             Connection connection = db.getConn();
             Statement statement = connection.createStatement();
             String query = "UPDATE coins SET coin = coin + " + coins + " WHERE id = '" + id + "'";
             statement.executeUpdate(query);
+            log("Add Coins");
             return "Successfully added coins";
         }catch (Exception e){
             e.printStackTrace();
@@ -72,12 +95,17 @@ public class CoinServicesImpl implements CoinServices{
 
     @WebMethod
     public String substractCoins(String id, int coins) {
+        if(!validateApiKey()){
+            System.out.println("API KEY INVALID");
+            return null;
+        }
         try{
             Database db = new Database();
             Connection connection = db.getConn();
             Statement statement = connection.createStatement();
             String query = "UPDATE coins SET coin = coin - " + coins + " WHERE id = '" + id + "'";
             statement.executeUpdate(query);
+            log("Substract Coins");
             return "Successfully removed coins";
         }catch (Exception e){
             e.printStackTrace();
@@ -87,18 +115,51 @@ public class CoinServicesImpl implements CoinServices{
 
     @WebMethod
     public String deleteCoins(String id) {
+        if(!validateApiKey()){
+            System.out.println("API KEY INVALID");
+            return null;
+        }
         try{
             Database db = new Database();
             Connection connection = db.getConn();
             Statement statement = connection.createStatement();
             String query = "DELETE FROM coins WHERE id = '" + id + "'";
             statement.executeUpdate(query);
+            log("Delete Coins");
             return "Successfully deleted coins";
         }catch (Exception e){
             e.printStackTrace();
         }
         return "Failed to delete coins";
     }
-
+    public Boolean validateApiKey() {
+        MessageContext messageContext = wsContext.getMessageContext();
+        String queryString = (String) messageContext.get("javax.xml.ws.http.request.querystring");
+        System.out.println("messageContext: " + queryString);
+        String[] keyValue = queryString.split("=");
+        if(keyValue.length==0 || !keyValue[0].equals("APIkey")){
+            return false;
+        }
+        String apiKey = keyValue[1];
+        System.out.println("API KEY: " + apiKey);
+        if(apiKey.equals("lalala")||apiKey.equals("hahaha")){
+            return true;
+        }else{
+            return false;
+        }
+    }
+    public void log(String description) {
+        MessageContext messageContext = wsContext.getMessageContext();
+        String queryString = (String) messageContext.get("javax.xml.ws.http.request.querystring");
+        System.out.println("messageContext: " + queryString);
+        String[] keyValue = queryString.split("=");
+        String apiKey = keyValue[1];
+        System.out.println("API KEY: " + apiKey);
+        String ip = "123";
+        String endpoint = (String) messageContext.get("javax.xml.ws.service.endpoint.address");
+        Timestamp timestamp = new Timestamp(new Date().getTime());
+        LogModel logModel = new LogModel();
+        logModel.InsertLog(description, endpoint, ip, timestamp.toString());
+    }
 
 }
